[CT-5-1] 리뷰반영 - SettingRepository 를 제거하고 UserRepository 로 통합
