+++ domain/src/main/java/com/yessorae/domain/usecase/ChangeCommissionRateSettingUseCase.kt
@@ -2,16 +2,16 @@ package com.yessorae.domain.usecase
 
 import com.yessorae.domain.common.Result
 import com.yessorae.domain.common.delegateEmptyResultFlow
-import com.yessorae.domain.repository.SettingRepository
+import com.yessorae.domain.repository.UserRepository
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 class ChangeCommissionRateSettingUseCase @Inject constructor(
-    private val settingRepository: SettingRepository
+    private val userRepository: UserRepository
 ) {
     operator fun invoke(rate: Double): Flow<Result<Unit>> =
         flow<Nothing> {
-            settingRepository.updateCommissionRate(rate)
+            userRepository.updateCommissionRate(rate)
         }.delegateEmptyResultFlow()
 }
