+++ domain/src/main/java/com/yessorae/domain/usecase/ChangeTickUnitSettingUseCase.kt
@@ -3,16 +3,16 @@ package com.yessorae.domain.usecase
 import com.yessorae.domain.common.Result
 import com.yessorae.domain.common.delegateEmptyResultFlow
 import com.yessorae.domain.entity.tick.TickUnit
-import com.yessorae.domain.repository.SettingRepository
+import com.yessorae.domain.repository.UserRepository
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 class ChangeTickUnitSettingUseCase @Inject constructor(
-    private val settingRepository: SettingRepository
+    private val userRepository: UserRepository
 ) {
     suspend operator fun invoke(tickUnit: TickUnit): Flow<Result<Unit>> =
         flow<Nothing> {
-            settingRepository.updateTickUnit(tickUnit)
+            userRepository.updateTickUnit(tickUnit)
         }.delegateEmptyResultFlow()
 }
