+++ domain/src/main/java/com/yessorae/domain/usecase/SubscribeChartGameUseCase.kt
@@ -5,20 +5,18 @@ import com.yessorae.domain.common.delegateValueResultFlow
 import com.yessorae.domain.entity.ChartGame
 import com.yessorae.domain.repository.ChartGameRepository
 import com.yessorae.domain.repository.ChartRepository
-import com.yessorae.domain.repository.SettingRepository
 import com.yessorae.domain.repository.UserRepository
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 
 class SubscribeChartGameUseCase @Inject constructor(
     private val userRepository: UserRepository,
-    private val settingRepository: SettingRepository,
     private val chartRepository: ChartRepository,
     private val chartGameRepository: ChartGameRepository
 ) {
     suspend operator fun invoke(gameId: Long?): Flow<Result<ChartGame>> {
         if (gameId == null) {
-            val totalTurn = settingRepository.fetchTotalTurn()
+            val totalTurn = userRepository.fetchTotalTurn()
 
             val newGameId = chartGameRepository.createNewChartGame(
                 chartGame = ChartGame.new(
