+++ domain/src/main/java/com/yessorae/domain/usecase/SubscribeCommissionRateSettingUseCase.kt
@@ -2,15 +2,15 @@ package com.yessorae.domain.usecase
 
 import com.yessorae.domain.common.Result
 import com.yessorae.domain.common.delegateValueResultFlow
-import com.yessorae.domain.repository.SettingRepository
+import com.yessorae.domain.repository.UserRepository
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 
 class SubscribeCommissionRateSettingUseCase @Inject constructor(
-    private val settingRepository: SettingRepository
+    private val userRepository: UserRepository
 ) {
     operator fun invoke(): Flow<Result<Double>> =
-        settingRepository
+        userRepository
             .fetchCommissionRateAsFlow()
             .delegateValueResultFlow()
 }
