+++ domain/src/main/java/com/yessorae/domain/usecase/SubscribeTickUnitSettingUseCase.kt
@@ -3,15 +3,15 @@ package com.yessorae.domain.usecase
 import com.yessorae.domain.common.Result
 import com.yessorae.domain.common.delegateValueResultFlow
 import com.yessorae.domain.entity.tick.TickUnit
-import com.yessorae.domain.repository.SettingRepository
+import com.yessorae.domain.repository.UserRepository
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 
 class SubscribeTickUnitSettingUseCase @Inject constructor(
-    private val settingRepository: SettingRepository
+    private val userRepository: UserRepository
 ) {
     operator fun invoke(): Flow<Result<TickUnit>> =
-        settingRepository
+        userRepository
             .fetchTickUnitAsFlow()
             .delegateValueResultFlow()
 }
